+++ packages/fe-fpm-writer/test/unit/building-block/prompts/utils/xml.test.ts
@@ -2,9 +2,9 @@ import { create as createStorage } from 'mem-fs';
 import type { Editor } from 'mem-fs-editor';
 import { create } from 'mem-fs-editor';
 import { join } from 'node:path';
-import { isElementIdAvailable, getOrAddNamespace } from '../../../../../src/building-block/prompts/utils/xml';
+import { getOrAddNamespace } from '../../../../../src/building-block/prompts/utils/xml';
 import { DOMParser } from '@xmldom/xmldom';
-import { BuildingBlockType } from '../../../../../src/building-block/types';
+import { isElementIdAvailable } from '../../../../../src/common/utils';
 
 describe('utils - xml', () => {
     let fs: Editor;