+++ packages/preview-middleware-client/src/cpe/additional-change-info/add-xml-additional-info.ts
@@ -1,5 +1,5 @@
 import FlexChange from 'sap/ui/fl/Change';
-import { getControlById, findViewByControl } from '../../utils/core';
+import { getControlBySelector, findViewByControl } from '../../utils/core';
 import {
     ANALYTICAL_TABLE_TYPE,
     GRID_TABLE_TYPE,
@@ -7,6 +7,8 @@ import {
     TREE_TABLE_TYPE
 } from '../../adp/quick-actions/control-types';
 import Element from 'sap/ui/core/Element';
+import type Component from 'sap/ui/core/Component';
+import type Selector from 'sap/ui/fl/Selector';
 
 export type AddXMLAdditionalInfo = {
     templateName?: string;
@@ -19,12 +21,15 @@ export type AddXMLChangeContent = {
     targetAggregation?: string;
 };
 
-export function getAddXMLAdditionalInfo(change: FlexChange<AddXMLChangeContent>): AddXMLAdditionalInfo | undefined {
-    const selectorId = change.getSelector()?.id ?? '';
+export function getAddXMLAdditionalInfo(
+    change: FlexChange<AddXMLChangeContent>,
+    appComponent?: Component
+): AddXMLAdditionalInfo | undefined {
+    const selector = change.getSelector();
     const targetAggregation = change.getContent()?.targetAggregation ?? '';
-    const targetControl = getControlById(selectorId);
+    const targetControl = getControlBySelector(selector, appComponent);
     const controlType = targetControl?.getMetadata().getName() ?? '';
-    const templateName = getFragmentTemplateName(selectorId, targetAggregation);
+    const templateName = getFragmentTemplateName(selector, targetAggregation, appComponent);
     const viewName = targetControl ? (findViewByControl(targetControl)?.getViewName() ?? '') : '';
 
     const result: AddXMLAdditionalInfo = {};
@@ -40,8 +45,12 @@ export function getAddXMLAdditionalInfo(change: FlexChange<AddXMLChangeContent>)
     return Object.keys(result).length > 0 ? result : undefined;
 }
 
-export function getFragmentTemplateName(selectorId: string, targetAggregation: string): string {
-    const control = getControlById(selectorId);
+export function getFragmentTemplateName(
+    selector: Selector | undefined,
+    targetAggregation: string,
+    appComponent?: Component
+): string {
+    const control = getControlBySelector(selector, appComponent);
 
     if (!control) {
         return '';