+++ packages/eslint-plugin-fiori-tools/CHANGELOG.md
@@ -1,5 +1,152 @@
 # @sap-ux/eslint-plugin-fiori-tools
 
+## 9.7.5
+
+### Patch Changes
+
+- Updated dependencies [d834713]
+    - @sap-ux/project-access@1.35.13
+    - @sap-ux/fiori-annotation-api@0.9.29
+
+## 9.7.4
+
+### Patch Changes
+
+- 2917c4c: fix(deps): update dependency yaml to v2.8.2
+- 7c06ef0: fix(deps): update dependencies [open-ux-odata]
+- fa74f92: fix(deps): update dependency c8 to v11
+- 83ca0e9: fix(deps): update dependency cross-env to v10
+- Updated dependencies [7c06ef0]
+    - @sap-ux/fiori-annotation-api@0.9.28
+    - @sap-ux/project-access@1.35.12
+
+## 9.7.3
+
+### Patch Changes
+
+- f90fcca: fix(deps): update dependency @eslint/json to v0.14.0
+    - @sap-ux/fiori-annotation-api@0.9.27
+    - @sap-ux/project-access@1.35.11
+
+## 9.7.2
+
+### Patch Changes
+
+- 4dcf980: fix(deps): update dependency @eslint/config-helpers to v0.5.2
+
+## 9.7.1
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.11
+- @sap-ux/fiori-annotation-api@0.9.26
+
+## 9.7.0
+
+### Minor Changes
+
+- 51236d3: Add sap-anchor-bar-visible eslint rule.
+
+## 9.6.10
+
+### Patch Changes
+
+- @sap-ux/fiori-annotation-api@0.9.25
+- @sap-ux/odata-annotation-core@0.2.16
+- @sap-ux/odata-entity-model@0.3.6
+- @sap-ux/odata-vocabularies@0.4.27
+
+## 9.6.9
+
+### Patch Changes
+
+- Updated dependencies [ca2566b]
+    - @sap-ux/project-access@1.35.10
+    - @sap-ux/fiori-annotation-api@0.9.24
+
+## 9.6.8
+
+### Patch Changes
+
+- 562469f: Update the readme with .mjs example for configuration
+
+## 9.6.7
+
+### Patch Changes
+
+- @sap-ux/fiori-annotation-api@0.9.23
+
+## 9.6.6
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.9
+- @sap-ux/fiori-annotation-api@0.9.22
+
+## 9.6.5
+
+### Patch Changes
+
+- Updated dependencies [0ecc5f1]
+- Updated dependencies [cc1c422]
+    - @sap-ux/project-access@1.35.8
+    - @sap-ux/fiori-annotation-api@0.9.21
+    - @sap-ux/odata-annotation-core@0.2.15
+    - @sap-ux/odata-entity-model@0.3.6
+    - @sap-ux/odata-vocabularies@0.4.26
+
+## 9.6.4
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.7
+- @sap-ux/fiori-annotation-api@0.9.20
+
+## 9.6.3
+
+### Patch Changes
+
+- @sap-ux/fiori-annotation-api@0.9.19
+
+## 9.6.2
+
+### Patch Changes
+
+- bb310dc: fix(deps): update dependency semver to v7.7.4
+- Updated dependencies [bb310dc]
+- Updated dependencies [e5bc3ca]
+    - @sap-ux/project-access@1.35.6
+    - @sap-ux/fiori-annotation-api@0.9.18
+
+## 9.6.1
+
+### Patch Changes
+
+- Updated dependencies [d57cc47]
+    - @sap-ux/project-access@1.35.5
+    - @sap-ux/fiori-annotation-api@0.9.17
+
+## 9.6.0
+
+### Minor Changes
+
+- 552c106: Fix: check table paste functionality is enabled on object pages only.
+
+## 9.5.11
+
+### Patch Changes
+
+- Updated dependencies [1fa3bb7]
+    - @sap-ux/odata-vocabularies@0.4.25
+    - @sap-ux/fiori-annotation-api@0.9.16
+
+## 9.5.10
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.4
+- @sap-ux/fiori-annotation-api@0.9.15
+
 ## 9.5.9
 
 ### Patch Changes