+++ packages/fiori-mcp-server/CHANGELOG.md
@@ -1,5 +1,93 @@
 # @sap-ux/fiori-mcp-server
 
+## 0.6.41
+
+### Patch Changes
+
+- 7c06ef0: fix(deps): update dependencies [open-ux-odata]
+    - @sap-ux/fiori-docs-embeddings@0.4.16
+
+## 0.6.40
+
+### Patch Changes
+
+- e773022: fix(deps): update dependency @sap-ai-sdk/foundation-models to v2.8.0
+    - @sap-ux/fiori-docs-embeddings@0.4.16
+    - @sap-ux/store@1.5.9
+
+## 0.6.39
+
+### Patch Changes
+
+- afd2fa2: feat: add icon and title to mcp server
+
+## 0.6.38
+
+### Patch Changes
+
+- a8f4e03: fix(deps): update dependency @sap-ai-sdk/langchain to v2.8.0
+- 2302698: fix(deps): update dependency @sap-ux/edmx-parser to v0.10.0
+    - @sap-ux/fiori-docs-embeddings@0.4.16
+
+## 0.6.37
+
+### Patch Changes
+
+- 4af92b5: add node: proto prefix to imports
+    - @sap-ux/fiori-docs-embeddings@0.4.16
+
+## 0.6.36
+
+### Patch Changes
+
+- Updated dependencies [ca2566b]
+    - @sap-ux/fiori-docs-embeddings@0.4.16
+
+## 0.6.35
+
+### Patch Changes
+
+- Updated dependencies [6c993f3]
+    - @sap-ux/store@1.5.8
+    - @sap-ux/fiori-docs-embeddings@0.4.15
+
+## 0.6.34
+
+### Patch Changes
+
+- 94d370f: fix(deps): update dependency @langchain/mcp-adapters to v1.1.3
+    - @sap-ux/fiori-docs-embeddings@0.4.15
+
+## 0.6.33
+
+### Patch Changes
+
+- ff634b0: fix(deps): update dependency @sap-ux/edmx-parser to v0.9.8
+    - @sap-ux/fiori-docs-embeddings@0.4.15
+
+## 0.6.32
+
+### Patch Changes
+
+- cbd340a: fix(deps): update dependency i18next to v25.8.12
+- Updated dependencies [cbd340a]
+    - @sap-ux/store@1.5.7
+    - @sap-ux/fiori-docs-embeddings@0.4.15
+
+## 0.6.31
+
+### Patch Changes
+
+- 1371f6b: fix(deps): update dependency @langchain/core to v1.1.26
+    - @sap-ux/fiori-docs-embeddings@0.4.15
+
+## 0.6.30
+
+### Patch Changes
+
+- Updated dependencies [d57cc47]
+    - @sap-ux/fiori-docs-embeddings@0.4.15
+
 ## 0.6.29
 
 ### Patch Changes